Move Core.ID macros to ID.Definition to avoid compile-time dependencies
