+++ helix/lib/game/ids/ids.ex
@@ -1,8 +1,8 @@
-defmodule Game.Connection.ID, do: use(Core.ID)
-defmodule Game.ConnectionGroup.ID, do: use(Core.ID)
-defmodule Game.File.ID, do: use(Core.ID)
-defmodule Game.Installation.ID, do: use(Core.ID)
-defmodule Game.Log.ID, do: use(Core.ID)
-defmodule Game.Process.ID, do: use(Core.ID)
-defmodule Game.Server.ID, do: use(Core.ID)
-defmodule Game.Tunnel.ID, do: use(Core.ID)
+defmodule Game.Connection.ID, do: use(Core.ID.Definition)
+defmodule Game.ConnectionGroup.ID, do: use(Core.ID.Definition)
+defmodule Game.File.ID, do: use(Core.ID.Definition)
+defmodule Game.Installation.ID, do: use(Core.ID.Definition)
+defmodule Game.Log.ID, do: use(Core.ID.Definition)
+defmodule Game.Process.ID, do: use(Core.ID.Definition)
+defmodule Game.Server.ID, do: use(Core.ID.Definition)
+defmodule Game.Tunnel.ID, do: use(Core.ID.Definition)
