+++ helix/lib/game/log/schemas/log.ex
@@ -34,13 +34,13 @@ defmodule Game.Log do
   ]
 
   @schema [
-    {:id, ID.ref(:log_id)},
+    {:id, ID.Definition.ref(:log_id)},
     {:revision_id, :integer},
     {:type, {:enum, values: @log_types}},
     {:direction, {:enum, values: @log_directions}},
     {:data, {:map, load_structs: true, after_read: :hydrate_data}},
     {:inserted_at, {:datetime_utc, [precision: :millisecond], mod: :inserted_at}},
-    {:server_id, {ID.ref(:server_id), virtual: true, after_read: :get_server_id}}
+    {:server_id, {ID.Definition.ref(:server_id), virtual: true, after_read: :get_server_id}}
   ]
 
   def new(params) do
