+++ helix/lib/game/network/schemas/network_connection.ex
@@ -6,7 +6,7 @@ defmodule Game.NetworkConnection do
 
   @schema [
     {:nip, Core.NIP},
-    {:server_id, ID.ref(:server_id)},
+    {:server_id, ID.Definition.ref(:server_id)},
     {:inserted_at, {:datetime_utc, [precision: :millisecond], mod: :inserted_at}}
   ]
