+++ helix/lib/game/software/schemas/installation.ex
@@ -14,13 +14,13 @@ defmodule Game.Installation do
   ]
 
   @schema [
-    {:id, ID.ref(:installation_id)},
+    {:id, ID.Definition.ref(:installation_id)},
     {:file_type, {:enum, values: @file_types}},
     {:file_version, :integer},
-    {:file_id, {ID.ref(:file_id), nullable: true}},
+    {:file_id, {ID.Definition.ref(:file_id), nullable: true}},
     {:memory_usage, :integer},
     {:inserted_at, {:datetime_utc, [precision: :millisecond], mod: :inserted_at}},
-    {:server_id, {ID.ref(:server_id), virtual: true, after_read: :get_server_id}}
+    {:server_id, {ID.Definition.ref(:server_id), virtual: true, after_read: :get_server_id}}
   ]
 
   @derived_fields [:id]
